Revert intcode back to ints

+++ src/y2019/day2/Part1.java
@@ -4,16 +4,16 @@ import utils.Problem;
 import y2019.intcode.Machine;
 import y2019.intcode.Parser;
 
-public class Part1 extends Problem<Long> {
+public class Part1 extends Problem<Integer> {
     private Parser parser;
 
-    public Part1(String fileName, long exp) {
+    public Part1(String fileName, int exp) {
         super(exp);
 
         parser = new Parser(fileName);
     }
 
-    public Long run() {
+    public Integer run() {
         var prog = parser.parse();
         var mach = new Machine(prog);
